backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
gitpython
python-dotenv
tiktoken
pydantic
pydantic-settings